requirements.txt
fastapi==0.115.0
uvicorn[standard]==0.30.6
httpx==0.27.2
pydantic==2.8.2
pydantic-settings==2.5.2
python-dotenv==1.0.1
stripe==12.4.0
pydantic[email]